backend/requirements.txt
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
backports.zoneinfo==0.2.1
cffi==1.17.1
channels==4.2.0
click==8.1.8
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==4.2.17
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
exceptiongroup==1.2.2
Faker==33.3.0
h11==0.14.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
psycopg==3.2.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.1
zope.interface==7.2
